fix(deps): resolve all dependency and import issues
- Fix package naming conflicts in testutil and webui
- Add Helm v3 dependencies for CNF orchestration
- Add missing replace directives for local packages
- Fix interface implementations in clients and config
- Add missing methods to secret managers
- Fix shared package imports
- Update K8s dependencies to v0.34.0
- Add cloud provider SDKs (AWS, Azure, GCP, VMware, OpenStack)
- Add security dependencies (SPIFFE, Vault, JWT)
- Add test frameworks (testcontainers, testify)
- Fix optimization/types.go merge conflicts
- Ensure all modules use consistent versions

+++ pkg/shared/rag_types.go
@@ -35,9 +35,9 @@ type SearchQuery struct {
 	UserID        string                 `json:"user_id,omitempty"`
 	Namespace     string                 `json:"namespace,omitempty"`
 	HybridSearch  bool                   `json:"hybrid_search,omitempty"`
-	HybridAlpha   float64                `json:"hybrid_alpha,omitempty"`
+	HybridAlpha   float32                `json:"hybrid_alpha,omitempty"`
 	UseReranker   bool                   `json:"use_reranker,omitempty"`
-	MinConfidence float64                `json:"min_confidence,omitempty"`
+	MinConfidence float32                `json:"min_confidence,omitempty"`
 	IncludeVector bool                   `json:"include_vector,omitempty"`
 	ExpandQuery   bool                   `json:"expand_query,omitempty"`
 	TargetVectors []string               `json:"target_vectors,omitempty"`
